Replace 'Linux' with 'GNU/Linux'
The term 'Linux' refers to a GNU environment. There are other
Linux-based systems, such as Android or Busybox. Ideally we'd
speak of a GNU environment, but readers might confuse this with
the full GNU system (e.g., HURD-based installations). The easiest
solution is to adopt the GNU/Linux term, which many readers are
familiar with.

Signed-off-by: Thomas Zimmermann <contact@tzimmermann.org>